requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
bcrypt==4.2.0
jinja2==3.1.6
httpx==0.28.1
email-validator==2.2.0
python-dateutil==2.9.0.post0
pypdf==5.6.0
python-docx==1.1.2
numpy==2.3.2
rapidfuzz==3.14.1
prometheus-client==0.23.1
pytest==8.4.1

